Added a Priority and Severity property to bug model.
Built a BugPriority and BugSeverity enum list.

Created a SetPriority method to convert an int value input into the equivalent in the enum list into a newvariable. To then set the new variable into the appropriate property; ie newPriority to Priority.

Created a SetServerity to do the same as SetPriority, but with the Severity property.

Built a BugService class to hold all sevices outside the Bug class.

Built a CreateBug Class that bulds a bug and adds to a list based on all the required inputs needed to fill out its propertiees.

Created tests to check that the CreateBug method bulds a bug as expected.

Created unit tests to check for SetPriority and SetSeverity that each instatiats correctly, their exceptions throw, and all property vaules display correctly.

Applyed some quality-of-life organization by creating regions for easy navigation and minimizing screen space usage.

Made some basic comments to help ease the understanding of the code.