Built UI intro menu for console app,
 Set it for the four variable choice of {Create Bug, View Bug List, Assign Develper, And Delete Bug}.
 Made it with a switch case so that it is easy to navigate.
 Set up for ease of unit testing by add as if statement to lock out Console.Clear() and Console.ReadKey()s.
Made a Ui for bug build competion,
Made Unit tests to check functionality of of Main Menu.
Insured that the Menu displays correctly,
Insured that When C was pressed, it was sucessful and went to desired location.
Made Unit test to insure that Bug Ticket displayed, and displayed correct information.